rename mocks to have different names (jest was complaining)
