+++ src/shared/utils/__tests__/getFASTAFromAccession.spec.ts
@@ -4,7 +4,7 @@ import MockAdapter from 'axios-mock-adapter';
 import getFASTAFromAccession from '../getFASTAFromAccession';
 
 import mockUniProtKB from '../../../uniprotkb/components/__mocks__/swissprotEntry.json';
-import mockUniParc from '../../../uniparc/__mocks__/entryModelData';
+import mockUniParc from '../../../uniparc/__mocks__/uniParcEntryModelData';
 // import mockUniRef from ...
 
 let mock;
