+++ src/tools/components/__tests__/SequenceSearchLoader.spec.tsx
@@ -4,7 +4,7 @@ import SequenceSearchLoader from '../SequenceSearchLoader';
 
 import customRender from '../../../shared/__test-helpers__/customRender';
 
-import entryModelData from '../../../uniprotkb/__mocks__/entryModelData';
+import entryModelData from '../../../uniprotkb/__mocks__/uniProtKBEntryModelData';
 import useDataApi from '../../../shared/hooks/useDataApi';
 
 jest.mock('../../../shared/hooks/useDataApi', () => jest.fn());
