+++ src/uniparc/__mocks__/uniParcEntryModelData.ts
