+++ src/uniprotkb/__mocks__/uniProtKBEntryModelData.ts
