+++ src/uniprotkb/adapters/__tests__/FamilyAndDomainsConverter.spec.tsx
@@ -1,7 +1,7 @@
 import convertFamilyAndDomains from '../familyAndDomainsConverter';
 import { convertXrefProperties } from '../uniProtkbConverter';
 
-import modelData from '../../__mocks__/entryModelData';
+import modelData from '../../__mocks__/uniProtKBEntryModelData';
 
 describe('Family and Domains data converter', () => {
   test('should convert the data', () => {
