+++ src/uniprotkb/adapters/__tests__/interactionConverter.spec.ts
@@ -1,7 +1,7 @@
 import convertInteraction from '../interactionConverter';
 import { convertXrefProperties } from '../uniProtkbConverter';
 
-import modelData from '../../__mocks__/entryModelData';
+import modelData from '../../__mocks__/uniProtKBEntryModelData';
 
 describe('Interaction data converter', () => {
   test('should convert the data', () => {
