+++ src/uniprotkb/adapters/__tests__/namesAndTaxonomyConverter.spec.ts
@@ -1,7 +1,7 @@
 import { convertNamesAndTaxonomy } from '../namesAndTaxonomyConverter';
 import { convertXrefProperties } from '../uniProtkbConverter';
 
-import modelData from '../../__mocks__/entryModelData';
+import modelData from '../../__mocks__/uniProtKBEntryModelData';
 
 describe('Names and taxonomy data converter', () => {
   test('should convert the data', () => {
