+++ src/uniprotkb/adapters/__tests__/structureConverter.spec.ts
@@ -1,7 +1,7 @@
 import convertStructure from '../structureConverter';
 import { convertXrefProperties } from '../uniProtkbConverter';
 
-import modelData from '../../__mocks__/entryModelData';
+import modelData from '../../__mocks__/uniProtKBEntryModelData';
 
 describe('Structure data converter', () => {
   test('should convert the data', () => {
