+++ src/uniprotkb/adapters/__tests__/subcellularLocationConverter.spec.ts
@@ -1,7 +1,7 @@
 import convertSubcellularLocation from '../subcellularLocationConverter';
 import { convertXrefProperties } from '../uniProtkbConverter';
 
-import modelData from '../../__mocks__/entryModelData';
+import modelData from '../../__mocks__/uniProtKBEntryModelData';
 
 describe('Subcellular data converter', () => {
   test('should convert the data', () => {
