+++ src/uniprotkb/components/entry/__tests__/EntryExternalLinks.spec.tsx
@@ -4,7 +4,7 @@ import { render } from '@testing-library/react';
 import EntryExternalLinks from '../EntryExternalLinks';
 import uniProtKbConverter from '../../../adapters/uniProtkbConverter';
 
-import mockData from '../../../__mocks__/entryModelData';
+import mockData from '../../../__mocks__/uniProtKBEntryModelData';
 
 describe('Entry - External Links view', () => {
   test('should render', () => {
